+++ src/main/java/com/health/yogiodigym/common/message/SuccessMessage.java
@@ -6,7 +6,8 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum SuccessMessage {
-    JOIN_SUCCESS("회원가입에 성공하였습니다.");
+    JOIN_SUCCESS("회원가입에 성공하였습니다."),
+    SEARCH_GYMS_SUCCESS("헬스장 조회에 성공하였습니다.");
 
     private final String message;
 }
