+++ src/main/java/com/health/yogiodigym/gym/service/GymService.java
@@ -1,33 +1,11 @@
 package com.health.yogiodigym.gym.service;
 
-import com.health.yogiodigym.gym.entity.DataGym;
-import com.health.yogiodigym.gym.repository.GymRepository;
+import com.health.yogiodigym.gym.dto.DataGymDto;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
 import java.util.Objects;
 
-@Service
-public class GymService {
+public interface GymService {
 
-    private final GymRepository gymRepository;
-
-    public GymService(GymRepository gymRepository) {
-        this.gymRepository = gymRepository;
-    }
-
-    public Page<DataGym> findByGymSearch(String gymKeyword, int page, int size, String searchColumn){
-        Pageable pageable = PageRequest.of(page, size);
-        if(!gymKeyword.isEmpty()){
-            if (Objects.equals(searchColumn, "oldAddress")){
-                return gymRepository.findByOldAddressContaining(gymKeyword, pageable);
-            }else {
-                return gymRepository.findByNameContaining(gymKeyword, pageable);
-            }
-        }else {
-            return gymRepository.findAll(pageable);
-        }
-    }
+    Page<DataGymDto> findByGymSearch(String gymKeyword, Pageable pageable, String searchColumn);
 }
