+++ src/main/java/com/health/yogiodigym/lesson/dto/CategoryDto.java
@@ -1,0 +1,22 @@
+package com.health.yogiodigym.lesson.dto;
+
+import com.health.yogiodigym.lesson.entity.Category;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CategoryDto {
+
+    private Long id;
+    private String name;
+    private String code;
+
+    public CategoryDto(Category category) {
+        this.id = category.getId();
+        this.name = category.getName();
+        this.code = category.getCode();
+    }
+}
