--- src/main/java/com/health/yogiodigym/lesson/dto/LessonEnrollmentDto.java
@@ -1,9 +1,0 @@
-package com.health.yogiodigym.lesson.dto;
-
-import lombok.Data;
-
-@Data
-public class LessonEnrollmentDto {
-    private Long memberId;
-    private Long lessonId;
-}
