+++ src/main/java/com/health/yogiodigym/lesson/entity/Category.java
@@ -9,6 +9,7 @@ import lombok.*;
 @Builder
 @Entity
 public class Category {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
