+++ src/main/java/com/health/yogiodigym/lesson/entity/Lesson.java
@@ -1,6 +1,6 @@
 package com.health.yogiodigym.lesson.entity;
 
-import com.health.yogiodigym.lesson.dto.LessonEditDto;
+import com.health.yogiodigym.lesson.dto.LessonDto;
 import com.health.yogiodigym.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.*;
@@ -15,6 +15,7 @@ import java.time.LocalTime;
 @Builder
 @Entity
 public class Lesson {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -61,7 +62,7 @@ public class Lesson {
         }
     }
 
-    public void updateLesson(LessonEditDto dto, Category category) {
+    public void updateLesson(LessonDto.Edit dto, Category category) {
         this.title = dto.getTitle();
         this.days = dto.getDays();
         this.category = category;
