+++ src/main/java/com/health/yogiodigym/lesson/repository/CategoryRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
+
     List<Category> findByCode(String code);
 }
