+++ src/main/java/com/health/yogiodigym/lesson/repository/LessonEnrollmentRepository.java
@@ -8,6 +8,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface LessonEnrollmentRepository extends JpaRepository<LessonEnrollment, Long> {
+
     boolean existsByLessonAndMember(Lesson lesson, Member member);
+
     Optional<LessonEnrollment> findByLessonAndMember(Lesson lesson, Member member);
 }
