--- src/main/resources/db/migration/V4__change_lesson.sql
@@ -1,5 +1,0 @@
-ALTER TABLE lesson
-    CHANGE day days INT NOT NULL;
-
-ALTER TABLE lesson_enrollment
-    DROP COLUMN unread_message;
