+++ src/main/resources/templates/lesson/register.html
@@ -7,6 +7,7 @@
     <script type="text/javascript" th:src="@{'//dapi.kakao.com/v2/maps/sdk.js?appkey=' + ${@environment.getProperty('kakao.api.scriptKey')} + '&libraries=services'}"></script>
     <script type="text/javascript" th:src="@{'/js/lessonRegister.js'}"></script>
     <script type="text/javascript" th:src="@{'/js/latLonApi.js'}"></script>
+    <script type="text/javascript" th:src="@{'/js/lessonValidateDate.js'}"></script>
 </head>
 <body>
 <h2>강의 등록</h2>
@@ -45,10 +46,10 @@
     <input type="time" name="endTime" required><br>
 
     <label>강의 시작일:</label>
-    <input type="date" name="startDay" required><br>
+    <input type="date" name="startDay" required onchange="validateEndDate()"><br>
 
     <label>강의 종료일:</label>
-    <input type="date" name="endDay" required><br>
+    <input type="date" name="endDay" required onchange="validateEndDate()"><br>
 
     <label>최대 인원:</label>
     <input type="number" name="max" required><br>
